Factor MockBuffer out
